Fixed a bug in the MultiHead Backpass
initially we did: grad = np.where(self.mask == 0, -1e9, grad)
This is wrong since we want to set the upper/lower triangular attention heads to -infinity before applying the softmax function within the forward pass to make the model autoregressive. exp(-infty) yields 0

In the backward pass we don't need to do this. This would lead to exploding gradients.
We need to just apply the logical mask of [0, 1] to the gradient.